requirements.txt
certifi==2021.10.8
click==8.0.1
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
minio==7.1.1
numpy==1.21.2
pandas==1.3.3
psycopg2==2.9.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
